Flake8 fixes.
