+++ python/pysmurf/client/command/smurf_command.py
@@ -15,6 +15,7 @@
 #-----------------------------------------------------------------------------
 import os
 import time
+import subprocess
 
 import numpy as np
 from packaging import version
@@ -6413,3 +6414,121 @@ class SmurfCommandMixin(SmurfBase):
         """
         return self._caget(self.stream_data_source + self._stream_data_period,
             **kwargs)
+
+    def shell_command(self,cmd,**kwargs):
+        r"""Runs command on shell and returns code, stdout, & stderr.
+
+        Args
+        ----
+        cmd : str
+            Command to run on shell.
+        \**kwargs
+            Arbitrary keyword arguments.  Passed directly to the
+            `subprocess.run` call.
+
+        Returns
+        -------
+        (stdout, stderr)
+            stdout and stderr returned as str
+        """
+        result = subprocess.run(
+            cmd.split(), stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE, shell=False, **kwargs
+        )
+
+        return result.stdout.decode(),result.stderr.decode()
+
+    def get_fru_info(self,board,bay=None,slot_number=None,shelf_manager=None):
+        r"""Returns FRU information for SMuRF board.
+
+        Wrapper for dumping the FRU information for SMuRF boards using
+        shell commands.
+
+        Args
+        ----
+        board : str
+            Which board to return FRU informationf for.  Valid options
+            include 'amc', 'carrier', or 'rtm'.  If 'amc', must also
+            provide the bay argument.
+        bay : int, optional, default None
+            Which bay to return the AMC FRU information for.  Used
+            only if board='amc'.
+        slot_number : int or None, optional, default None
+            The crate slot number that the AMC is installed into.  If
+            None, defaults to the
+            :class:`~pysmurf.client.base.smurf_control.SmurfControl`
+            class attribute
+            :attr:`~pysmurf.client.base.smurf_control.SmurfControl.slot_number`.
+        shelf_manager : str or None, optional, default None
+            Shelf manager ip address.  If None, defaults to the
+            :class:`~pysmurf.client.base.smurf_control.SmurfControl`
+            class attribute
+            :attr:`~pysmurf.client.base.smurf_control.SmurfControl.shelf_manager`.
+            For typical systems the default name of the shelf manager
+            is 'shm-smrf-sp01'.
+
+        Returns
+        -------
+        fru_info_dict : dict
+            Dictionary of requested FRU information.  Returns None if
+            board not a valid option, board not present in slot, slot
+            not present in shelf, or if no AMC is up in the requested
+            bay.
+        """
+        if slot_number is None:
+            slot_number=self.slot_number
+        if shelf_manager is None:
+            shelf_manager=self.shelf_manager
+
+        valid_board_options=['amc','rtm','carrier']
+        if board not in valid_board_options:
+            self.log(f'ERROR : {board} not in list of valid board options {valid_board_options}.  Returning None.',self.LOG_ERROR)
+            return None
+
+        shell_cmd=''
+        shell_cmd_prefix=None
+        if board=='amc':
+            shell_cmd_prefix='amc'
+            # require bay argument
+            if bay is None:
+                self.log('ERROR : Must provide AMC bay.  Returning None.',self.LOG_ERROR)
+                return None
+            if bay not in [0,1]:
+                self.log('ERROR : bay argument can only be 0 or 1.  Returning None.',self.LOG_ERROR)
+                return None
+            shell_cmd+=f'/{bay*2}'
+        elif board=='rtm':
+            # require bay argument
+            shell_cmd_prefix='rtm'
+        else: # only carrier left
+            shell_cmd_prefix='fru'
+
+        shell_cmd=f'cba_{shell_cmd_prefix}_init -d {shelf_manager}/{slot_number}'+shell_cmd
+        stdout,stderr=self.shell_command(shell_cmd)
+
+        # Error handling
+        if 'AMC not present in bay' in stdout:
+            self.log('ERROR : AMC not present in bay!  Returning None.',
+                     self.LOG_ERROR)
+            return None
+        if 'Slot not present in shelf' in stdout:
+            self.log('ERROR : Slot not present in shelf!  Returning None.',
+                     self.LOG_ERROR)
+            return None
+        if 'Board not present in slot' in stdout:
+            self.log('ERROR : Board not present in slot!  Returning None.',
+                     self.LOG_ERROR)
+            return None
+        else: # parse and return fru information for this board
+            stdout=stdout.split('\n')
+            fru_info_dict={}
+            for line in stdout:
+                if ':' in line:
+                    splitline=line.split(':')
+                    if len(splitline)==2:
+                        fru_key=splitline[0].lstrip().rstrip()
+                        fru_value=splitline[1].lstrip().rstrip()
+                        if len(fru_value)>0: # skip header
+                            fru_info_dict[fru_key]=fru_value
+
+        return fru_info_dict
