+++ python/pysmurf/client/util/smurf_util.py
@@ -4435,9 +4435,10 @@ class SmurfUtilMixin(SmurfBase):
                 plt.close()
 
         return gap_freq
-    
+
     def check_full_band_resp(self,
                              n_scan_per_band=5,
+                             amc_auto_detect=True,
                              make_plot=True,
                              save_plot=True,
                              show_plot=False,
@@ -4455,6 +4456,9 @@ class SmurfUtilMixin(SmurfBase):
         n_scan : int, optional, default 5
             The number of times to measure each band using
             full_band_resp.
+        amc_auto_detect : bool, optional, default True
+            If True, determines AMC RF bands from serial number.
+            Otherwise, uses `get_band_center_mhz`.
         make_plot : bool, optional, default True
             Whether to make a plot.
         save_plot : bool, optional, default True
@@ -4476,73 +4480,91 @@ class SmurfUtilMixin(SmurfBase):
         timestamp = self.get_timestamp()
         # Take which bands are configured from pysmurf configuration
         # file.
-        bands = self._bands
+        bands=self._bands
+
+        carrier_sn=self.get_carrier_sn(use_shell=True)
 
         results_dict={}
+        results_dict['n_scan_per_band']=n_scan_per_band
+        results_dict['amc_auto_detect']=amc_auto_detect
         results_dict['kwargs']=kwargs
-        for band in bands:
-            print(' ')
-            print(' ')
-            print(f'Band {band}')
-            print(' ')
-            print(' ')
-            results_dict[band]={}
-            results_dict[band]['fc']=self.get_band_center_mhz(band)
+        for bay in self.bays:
+            results_dict[bay]={}
+            bands=range(4*bay,4*(bay+1))
+            for band in bands:
+                print(' ')
+                print(' ')
+                print(f'Band {band}')
+                print(' ')
+                print(' ')
+                results_dict[bay][band%4]={}
+                amc_sn=self.get_amc_sn(bay=bay,use_shell=True)
+                results_dict[bay]['amc_sn']=amc_sn
+                amc_type=amc_sn.split('-')[1]
 
-            f,resp=self.full_band_resp(band=band, make_plot=False, show_plot=False, n_scan=n_scan_per_band, timestamp=timestamp, save_data=False, **kwargs)
-            results_dict[band]['f']=f
-            results_dict[band]['resp']=resp
+                # Sometimes the rogue zip files don't properly set the
+                # correct band center frequency if e.g. a LB is
+                # plugged into bay 1.
+                if amc_auto_detect and amc_type in ['A01','A02']:
+                    if amc_type=='A01': #LB
+                        results_dict[bay][band%4]['fc']=4250.+500.*(band%4)
+                    if amc_type=='A02': #HB
+                        results_dict[bay][band%4]['fc']=6250.+500.*(band%4)
 
-            # Also record the UC and DC attenuator settings for this
-            # band
-            results_dict[band]['uc_att']=self.get_att_uc(band)
-            results_dict[band]['dc_att']=self.get_att_dc(band)
+                else:
+                    results_dict[bay][band%4]['fc']=self.get_band_center_mhz(band)
 
-        # Plot results (if desired)
-        fig, ax = plt.subplots(2, figsize=(6,7.5), sharex=True)
-        try:
-            plt.suptitle(f'slot={self.slot_number} AMC0={self.get_amc_sn(0,use_shell=True)} AMC2={self.get_amc_sn(1,use_shell=True)}')
-        except:
-            # For some reason, FRU info wasn't accessible.  Give up.
-            plt.suptitle(f'slot={self.slot_number}')
+                f,resp=self.full_band_resp(band=band, make_plot=False, show_plot=False, n_scan=n_scan_per_band, timestamp=timestamp, save_data=False, **kwargs)
+                results_dict[bay][band%4]['f']=f
+                results_dict[bay][band%4]['resp']=resp
 
-        save_name = f'{timestamp}_full_band_resp_all.png'
-        ax[0].set_title(save_name)        
+                # Also record the UC and DC attenuator settings for this
+                # band
+                results_dict[bay][band%4]['uc_att']=self.get_att_uc(band)
+                results_dict[bay][band%4]['dc_att']=self.get_att_dc(band)
 
-        # Compute unwrapped phase for each band
-        last_angle=None
-        for band in bands:
-            f_plot=results_dict[band]['f']/1e6
-            resp_plot=results_dict[band]['resp']
-            plot_idx = np.where(np.logical_and(f_plot>-250, f_plot<250))
-            ax[0].plot(f_plot[plot_idx]+results_dict[band]['fc'], np.log10(np.abs(resp_plot[plot_idx])),label=f'b{band}')
-            angle = np.unwrap(np.angle(resp_plot))
-            if last_angle is not None:
-                angle-=(angle[0]-last_angle)
-            ax[1].plot(f_plot[plot_idx]+results_dict[band]['fc'], angle[plot_idx],label=f'b{band}')
-            results_dict[band]['phase']=angle[plot_idx]
-            last_angle=angle[plot_idx][-1]
+            # Plot results (if desired)
+            fig, ax = plt.subplots(2, figsize=(6,7.5), sharex=True)
+            plt.suptitle(f'slot={self.slot_number} / Carrier={carrier_sn} / AMC{bay*2}={amc_sn}')
 
-        ax[0].legend(loc='lower left',fontsize=8)
-        ax[0].set_ylabel("log10(abs(Response))")
-        ax[0].set_xlabel('Frequency [MHz]')
+            save_name = f'{timestamp}_full_band_resp_all_amc{bay*2}.png'
+            ax[0].set_title(save_name)
 
-        ax[1].legend(loc='lower left',fontsize=8)
-        ax[1].set_ylabel("Phase [rad]")
-        ax[1].set_xlabel('Frequency [MHz]')
+            # Compute unwrapped phase for each band
+            last_angle=None
+            for band in bands:
+                f_plot=results_dict[bay][band%4]['f']/1e6
+                resp_plot=results_dict[bay][band%4]['resp']
+                plot_idx = np.where(np.logical_and(f_plot>-250, f_plot<250))
+                ax[0].plot(f_plot[plot_idx]+results_dict[bay][band%4]['fc'], np.log10(np.abs(resp_plot[plot_idx])),label=f'b{band%4}')
+                angle = np.unwrap(np.angle(resp_plot))
+                if last_angle is not None:
+                    angle-=(angle[0]-last_angle)
+                ax[1].plot(f_plot[plot_idx]+results_dict[bay][band%4]['fc'], angle[plot_idx],label=f'b{band%4}')
+                results_dict[bay][band%4]['phase']=angle[plot_idx]
+                last_angle=angle[plot_idx][-1]
 
-        plt.tight_layout()
+            ax[0].legend(loc='lower left',fontsize=8)
+            ax[0].set_ylabel("log10(abs(Response))")
+            ax[0].set_xlabel('Frequency [MHz]')
+            ax[0].set_ylim(-0.5,1.25)
 
-        if save_plot:
-            save_path = os.path.join(self.plot_dir, save_name)
-            print(f'Saving plot to {save_path}.')
-            plt.savefig(save_path,
-                        bbox_inches='tight')
-            self.pub.register_file(save_path, 'fbr_plot', plot=True)
+            ax[1].legend(loc='lower left',fontsize=8)
+            ax[1].set_ylabel("Phase [rad]")
+            ax[1].set_xlabel('Frequency [MHz]')
 
-            if not show_plot:
-                plt.close()
-                
+            plt.tight_layout()
+
+            if save_plot:
+                save_path = os.path.join(self.plot_dir, save_name)
+                print(f'Saving plot to {save_path}.')
+                plt.savefig(save_path,
+                            bbox_inches='tight')
+                self.pub.register_file(save_path, f'fbr{bay*2}_plot', plot=True)
+
+                if not show_plot:
+                    plt.close()
+
         if save_results:
             save_path = os.path.join(self.output_dir,f'{timestamp}_full_band_resp_all.npy')
             print(f'Saving data to {save_path}.')
@@ -4551,5 +4573,5 @@ class SmurfUtilMixin(SmurfBase):
 
         if show_plot:
             plt.show()
-            
+
         return results_dict
